refactor: extract xref stream parsing from xref.go into xref_stream.go
